Improve Event Navigation in History (#42)
* **Added Static TOC Menu Bar**
When `menubar_toc: true` is added to the frontmatter for a page, there's
no static content. For example, from an event results page, there's no
easy way to get back to the season page. This change introduces a
`menubar_toc_static` `page` variable that accepts YAML directly defined
in the frontmatter for each page.

* **Added Season Links to All Events**
Updated the results for all events to include season links.