Add manual indexing for memories and older nationals (#206)
* **Memories**
The individual quizzer memories/recollections weren't tagged with the
`quizzerIndex`, which meant they weren't picked up during indexing. This
change adds manual indexing for those files.

* **Older Nationals**
Many of the older `nationals.md` didn't include all the quizzers, but
there was a list of the individual team members on the page. For pages
in this situation, this change creates a single `quizzerIndex` at the
top of the page to include any quizzers not listed in individual scores.