Update the quizzerIndex to only include unique quizzers (#208)
* **Quizzer Index Front Matter**
The `quizzerIndex` in the front matter only needs to include the
*unique* quizzer names that won't be discovered otherwise. This change
removes any quizzers that are discovered through regular indexing to
reduce the size of the pages.

* **Page Count**
The current quizzer search exposes an event count, but this ignores
pages without seasons. Instead, in this change, the label is now page
count.