Update source of live event scoring data (#270)
* **Moved Source of Live Event Data**
The back-end has been updated to use a different source
(scores.biblequiz.com) to allow for better caching and performance.
Additional changes are coming related to this.

* **Update TBQ Live Events**
The TBQ Live Events has a link to the 2023 Nationals rather than the
menu bar used in JBQ Live Events and the TBQ Current Season. This change
fixes the issue.

* **Handle Missing Meets**
If there's a link to an unknown database, it will now get a better error
message.