Fix Formatting Issues causing tables to not render correctly. (#324)
* **Win/Loss Columns**
The `W / L` columns are actually being rendered as two separate columns,
but the header has them combined. This change does a manual update for
the existing generated scores. There's a separate change going in to the
bot to avoid this altogether.

* **Spacing after Tables**
When an Individuals table doesn't have any ties, the header for the next
event appears on the very next line, which won't be formatted correctly.