Add Room and Grid Schedule View to Live Scores (#391)
## Issue
The updated Schedules tab in live scores was designed to display the
scores in a mobile friendly way. The old grid system didn't print well
and was targeted at the event coordinator. Unfortunately, the existing
system forces the user to go to the registration site to look at the
database settings.

## Fix
Introduced a new View button on the page:

![image](https://github.com/biblequiz/BibleQuiz.com/assets/15827689/1e7d6664-a849-454d-ac94-d20139c3b4bb)

This enables several new scenarios:
* **Room View**
Officials often want to see what teams are coming to their room next.
The existing view is team-centric, which makes this more difficult. This
change introduces a view pivoted by room. Additional improvements in the
future may make it easier to determine which room is delaying the match.

* **Grid View**
Displays the old view (currently accessible from the database settings
page) to allow event coordinators to more easily print and see the
dashboard for an event.

## Other Fixes
The change also includes some fixes to the loading bar which wasn't
being displayed correctly.