Update chapter locations for Festus in portion (#878)
The chapter locations mentioned for Festus in the Full Chapter Analysis
updates to the portion were incorrect. This corrects the page on
biblequiz.com. The newsletter PDFs should also be updated going forward.
I see they are PDFs so not sure how those get generated and uploaded to
the repo.