Add Prefix for Single Team Schedule or Scores (#1043)
When printing single team schedule or scores, the meet name doesn't
appear on any other page than the first. This can make it difficult for
the coordinator to keep track of the church's division when handing out
schedules. This change prefixes the team name with the name of the event
in this scenario.