Apply multiple fixes on the event page discovered during testing (#1415)
* **Hide Content when printing**
This change hides the BibleQuiz.com logo and status alert banner.

* **Add Status to Collapsible Section**
Updated the `CollapsableMeetSection` to include the status when viewing
meets in collapsed state. This also replaces the soccer ball icon for
events with the current status. This makes it simple to view the overall
status in a single view without expanding each event. This also includes
the new combined status exposed by the service when displaying linked
meet's status in a single row.

* **Set Combined Name**
The site isn't using the combined name when displaying room-level
information.