Re-Enable BibleQuiz.com Login (#1435)
* **Introduced the Sign Up Flow**
Added the sign-up flow for BibleQuiz.com users with integration to
Registration.BibleQuiz.com. The result is that a user created on
BibleQuiz.com will work with Registration.BibleQuiz.com. This creates
the foundation for merging apps into BibleQuiz.com with react.

* **Added Legal Language**
Brought the Privacy Policy and Terms & Conditions from
Registration.BibleQuiz.com to BibleQuiz.com.