Add the Question Generator and Refactor Logged In Experience (#1452)
* **Introduces the Question Generator**
Implements the Question Generator from registration.biblequiz.com, using
a similar UI, with full functional parity. The implementation will serve
as the example for how apps or widgets will be added for other sections
(e.g., event management).

* **Protected Routes**
If one of the pages requires auth, it is simply wrapped in a protected
route, which will display the sign-up prompt to the user with a friendly
message. This reduces the number of dialogs that get displayed. The auth
button is only displayed when the user is logged in.

* **Better Error Handling**
Implemented better error handling across the react components.