Fix Multiple Question Generator Bugs (#1459)
* **Move to String Serialized Enums**
Update the enums to be strings to avoid confusion, since they are now
serialized as strings in the backend.

* **Hide Point Value Rules for Manual Ordering**
If the user is manually ordering the point values, the point value rules
don't make sense.

* **Display Alert when downloading**
Downloading the PDF can take a little bit for larger sets. This change
adds an alert indicating the site is actually doing something.

* **Keep Alerts for 30 Seconds**
If there is an error, the alert will now appear for 30 seconds instead
of 5, making it easier to read and understand. The alert will be cleared
if the user starts another operation.