Improve Experience for Favorites (#1486)
* **Rename Favorites to Follow**
Changed the name of the Favorites functionality to "Follow", adding text
to make it clear what the functionality does.

* **Only Require Sign-in on Protected Pages**
The user will be requested to sign-in when they are navigating the web
site, even if the page isn't a protected one. This change moves to only
request this from the user when they are on a protected page.