Handle Reused Events (#1587)
If the dates for an event already published to the archive are moved
into the future, the new scores will not appear on the web site until
after an archive sync. As a performance improvement, the event page
prefers the file in the archive despite the dates. This change updates
it to consider the end date of the event before checking for the
existence of the file.