Fix in-room score links (#1608)
* **Always Add the Event Handler**
For unknown reasons, the `useEffect` in `RoomDialogLink` didn't always
get applied. It was added to support the scenario where the component is
rendered server-side, but it doesn't appear to function in those
scenarios. This change moves the event handler directly to the element
and this appears to function as expected.

* **Clean-up Naming**