Home Page, Results Page and Registration Page updates/fixes (#1624)
1. Updated the Home Page to include the Menu and the main JBQ and TBQ
buttons on top so that people can quickly discover their stuff and go
there. Pushed the LiveEvents down one level so its still right in front
of you but because its dynamic and grow and shrink, the real-estate is
properly used. I feel its more intuitive and functional with less
scrolling and button presses.
2. Updated the Results page to be more flexible and mobile friendly and
the scores and registration buttons to appear in single column on
mobile. This was a confusing one. Most of us are on mobile and we're not
able to see the scores until accidentally we scroll right. This is now
flexible div rendering instead of table so its more mobile friendly.
3. Finally, I could not get the RichTextEditor to work but also I see
that its not free. So updated with another github package thats light
weight and free for commercial use and works. Not sure if I configured
it right to read the text. Please check.

4. Updated highlight feature to be more dark and light mode friendly

---------

Co-authored-by: michsco <scottm@michaelbscott.com>