Address first round of feedback (#1684)
* **Hide Header when printing**
Hid the search box (#1677) and event title (#1682) when printing the
page as this is just using up space. Furthermore, the title and search
only appears on the first page. Subsequent events are missing the data.

* **Fixed Close Button on Print Dialog**
The `Close` button wasn't working correctly due to a missing event
handler. This change adds the missing event handler.

* **Fixed Cancel Button on Print Dialog**
The `Cancel` button on the browser's print dialog caused an error to
occur as the page attempted to remove the `QStats` tab a second time.
This change addresses #1674 by hiding the element instead of removing
it. This allows the code to be repeatable and avoids the error.

* **Moved Current Season**
The `Current Season` was confusing (#1671) in its current location as it
causes users to believe it is already expanded. This change moves the
TBQ version of `Current Season` to appear above `Next Season` (same
behavior as JBQ).

* **Fixed Print Dialog Meet Filter**
The meets selected in the Print Dialog on the web site aren't honored
when printing a page. This change addresses #1681 by honoring the
settings.