Add fixes for bugs (#1693)
* **Add Redirects for ScoreKeep QR Code**
ScoreKeep includes a QR Code on the printed schedule to
[biblequiz.com/jbq/live-events/](http://biblequiz.com/jbq/live-events/)
or
[biblequiz.com/jbq/live-events/](http://biblequiz.com/tbq/live-events/),
which don't exist on the web site. This fix for #1692 adds a Cloudflare
redirect to the Upcoming & Live Events with a `?type=<type>` URL
parameter that is handled by the page.

* **Add Location to Event Card**
Introduced the location to the Event Card so it can be used in the past
events.

* **Add Event Cards to Past Events**
Changed the table of past events to cards from the home page as this
renders better across different devices (#1672).

* **Enable 3 Columns on Home Page**
Shrunk the size of the Event Cards to allow 3 full columns for events to
be rendered.

* **Add Register Button to Event**
Added a Register button to the page for each event (if it is open).

* **Made Stats First**
Moved the `Stats` tab to the beginning of the list (#1675)