Update Event Lists on Home Page and Results Pages (#1793)
* **Added Filters to Home Page**
Introduced filters on the home page that reuse the same capabilities as
the Season pages. The filters are persisted across all pages to make the
experience cleaner. This new change reduces the number of cards for a
given user (assuming they use the filters) to a manageable amount.

* **Added Recent Events**
If a user checks the home page the day after a meet, it will disappear
from the home page. This change has a "Just Happened" section to display
recent events.

* **Add Logo for Pages**
Add the `eventType` logo (e.g., JBQ or TBQ) based on the URL (if it
hasn't already been set). This will make it more obvious when the user
is on a specific section of the web site.

* **Deleted Upcoming Events**
Removed the `/upcoming-events` in favor of the upcoming events on the
home page.

* **Removed Layouts**
The layouts directory isn't used.